refactor: 의도를 잘 모르는 쿼리 옵션 삭제 및 폴더 구조 일관성 맞추기
