+++ src/features/search-posts/model/useCategoriesQuery.ts
